requirements.txt
python-telegram-bot[job-queue]==21.9
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
asyncpg==0.29.0
aiosqlite==0.20.0